Duration of Each Event

This function calculates the duration of each event of collective motion in a dataset.

Usage

calc_dur_per_event(data, step2time)

Arguments

data

A dataframe with an event column (added by get_event_ids), indicating the event ID that each timestep belongs to. Timesteps that are not part of an event should not be included in the data.

step2time

The sampling frequency of the dataframe (how many seconds are between each row of the data).

Value

A dataframe with two columns, event ID and duration in seconds.

See also

Author

Marina Papadopoulou m.papadopoulou.rug@gmail.com

Examples

data <- data.frame(
set = c(rep('1', 50), rep('2', 50)),
event = c(rep(NA, 10), rep(1, 40), rep(2, 30), rep(NA, 20))
)

time_per_row <- 1 # seconds

calc_dur_per_event(data, time_per_row)
#>   event event_dur
#> 1     1        40
#> 2     2        30